Login&Registration:Both UI has been integrated on same screen
code refactoring

Widget created for login and registration
